// src/openapi_p/models/EpisodeWikiInfo.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { EpisodeType } from './EpisodeType';
export type EpisodeWikiInfo = {
    type: EpisodeType;
    /**
     * YYYY-MM-DD
     */
    date?: string;
    disc?: number;
    duration: string;
    ep: number;
    id: number;
    name: string;
    nameCN: string;
    subjectID: number;
    summary: string;
};