// src/openapi_p/models/PersonWikiInfo.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SubjectType } from './SubjectType';
export type PersonWikiInfo = {
    typeID: SubjectType;
    id: number;
    infobox: string;
    name: string;
    summary: string;
};