// src/openapi_p/models/RecentWikiChange.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RecentWikiChange = {
    persons: Array<{
        createdAt: number;
        id: number;
    }>;
    subject: Array<{
        createdAt: number;
        id: number;
    }>;
};