// src/openapi_p/models/SubjectEdit.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SubjectEdit = {
    date?: string;
    infobox: string;
    name: string;
    nsfw: boolean;
    platform: number;
    summary: string;
    metaTags: Array<string>;
};